+++ lib/formBuilder/styles.d.ts
@@ -1,14 +1,14 @@
 export declare const arrows: {
-    '& .arrows': {
-        float: string;
-        '& .fa-arrow-up, & .fa-arrow-down': {
-            'border-radius': string;
-            padding: string;
-            margin: string;
-            border: string;
-            color: string;
-            height: string;
-            width: string;
+    readonly '& .arrows': {
+        readonly float: "right";
+        readonly '& .fa-arrow-up, & .fa-arrow-down': {
+            readonly borderRadius: "4px";
+            readonly padding: ".25em";
+            readonly margin: "0 .5em 0 0";
+            readonly border: "1px solid #1d71ad";
+            readonly color: "#1d71ad";
+            readonly height: "28px";
+            readonly width: "28px";
         };
     };
 };