+++ lib/textFieldContext/TextField.d.ts
@@ -1,4 +1,4 @@
 import React from 'react';
-import { TextFieldProps } from '@material-ui/core/TextField';
+import { TextFieldProps } from '@mui/material/TextField';
 declare const TextField: (props: TextFieldProps) => React.JSX.Element;
 export default TextField;